move haversine test to new script
