+++ tests/test_comparison_level_lib.py
@@ -2,7 +2,6 @@ import pandas as pd
 
 from splink.duckdb.duckdb_comparison_level_library import (
     columns_reversed_level,
-    distance_in_km_level,
     else_level,
     exact_match_level,
     null_level,
@@ -48,70 +47,6 @@ def test_column_reversal():
 
     row = dict(df_e.query("id_l == 3 and id_r == 4").iloc[0])
     assert row["gamma_full_name"] == 2
-
-
-def test_haversine_level():
-    data = [
-        {"id": 1, "lat": 22.730590, "lon": 9.388589},
-        {"id": 2, "lat": 22.836322, "lon": 9.276112},
-        {"id": 3, "lat": 37.770850, "lon": 95.689880},
-        {"id": 4, "lat": -31.336319, "lon": 145.183685},
-    ]
-    # Add another the array version of the lat_long column
-    for d in data:
-        d["lat_long"] = {"lat": d["lat"], "long": d["lon"]}
-        d["lat_long_arr"] = [d["lat"], d["lon"]]
-
-    df = pd.DataFrame(data)
-
-    settings = {
-        "unique_id_column_name": "id",
-        "link_type": "dedupe_only",
-        "blocking_rules_to_generate_predictions": [],
-        "comparisons": [
-            {
-                "output_column_name": "lat_long",
-                "comparison_levels": [
-                    null_level("lat"),  # no nulls in test data
-                    distance_in_km_level(
-                        km_threshold=50,
-                        lat_col="lat",
-                        long_col="lon",
-                    ),
-                    distance_in_km_level(
-                        lat_col="lat_long['lat']",
-                        long_col="lat_long['long']",
-                        km_threshold=10000,
-                    ),
-                    distance_in_km_level(
-                        lat_col="lat_long_arr[1]",
-                        long_col="lat_long_arr[2]",
-                        km_threshold=100000,
-                    ),
-                    else_level(),
-                ],
-            },
-        ],
-        "retain_matching_columns": True,
-        "retain_intermediate_calculation_columns": True,
-    }
-
-    linker = DuckDBLinker(df, settings, input_table_aliases="test")
-    df_e = linker.predict().as_pandas_dataframe()
-
-    row = dict(df_e.query("id_l == 1 and id_r == 2").iloc[0])
-    assert row["gamma_lat_long"] == 3
-
-    # id comparisons w/ dist < 10000km
-    id_comb = {(1, 3), (2, 3), (3, 4)}
-    for id_pair in id_comb:
-        row = dict(df_e.query("id_l == {} and id_r == {}".format(*id_pair)).iloc[0])
-        assert row["gamma_lat_long"] == 2
-
-    id_comb = {(1, 4), (2, 4)}
-    for id_pair in id_comb:
-        row = dict(df_e.query("id_l == {} and id_r == {}".format(*id_pair)).iloc[0])
-        assert row["gamma_lat_long"] == 1
 
 
 def test_perc_difference():
